refactor: primitiveColors를 @repo/design-tokens 패키지로 분리
