+++ packages/design-tokens/src/index.ts
@@ -1,0 +1,1 @@
+export { primitiveColors, type PrimitiveColors } from './colors';
